Squashed commit of the following:
commit fa54e2b81e45a2cd815068e037ab05f9e785cd78
Author: Moe <github@m03.ca>
Date:   Tue May 7 23:43:16 2024 -0700

    make update on container start optional

commit 873df5e0ad1af002156658b0424e41a01fc16063
Author: Moe <github@m03.ca>
Date:   Tue May 7 23:30:49 2024 -0700

    remove depending on shinobi-sql

+++ entrypoint.sh
@@ -34,8 +34,13 @@ EOF
 )
 echo "Setting Database"
 echo $DB_CONFIG
-git reset --hard
-git pull
+
+if [ "$SHINOBI_UPDATE" = "true" ]; then
+    echo "Updating Shinobi..."
+    git reset --hard
+    git pull --rebase
+fi
+
 node tools/modifyConfiguration.js addToConfig="{\"db\": $DB_CONFIG}"
 
 pm2 flush
